after AOA break + code revision pending

+++ src/Crio/DSA_3/ZigzagLevelOrder.java
@@ -13,7 +13,7 @@ public class ZigzagLevelOrder {
             int size = queue.size();
             for(int i=0;i<size;i++){
                 TreeNode curr = queue.poll();
-                subLst.add(curr.val);
+                subLst.add((long) curr.val);
                 if(curr.left!=null) queue.add(curr.left);
                 if(curr.right!=null)    queue.add(curr.right);
             }
